adds --exec_script flag to run pre-build command
Now sphinx-autobuild has the ability to run a script before building.

If the script errors out, sphinx-autobuild will print out a warning, but
otherwise continue.

My use case for this feature:
In my previous workflow of manually rebuilding, I had a script that
would automatically add in an `automodule` directive in the test
documentation. If someone changes what a test script references, I want
the `automodule` directives to also update.