ioloop: Handle multiple file changes at once during live view
When one multiple files within a repository at once (e.g. changing
branches for instance), the live view will cause multiple consecutive
rebuilds of the documentation set even though it already had built
against the new changes.  This is especially visible when building the
PDF variants of the documentation.

This change works well if pyinotify is installed; without pyinotify,
this behaves more or less like it did previously.  The livereload
integration with pyinotify will invoke __call__(self) multiple times in
a row with all of the changed files and then the top-level ioloop will
invoke our callback once for all of the changed files.