[fix] interaction with autodoc plugin / sphinx-build in subprocess
To update changes to the doc strings (autodoc) in the source code in the HTML
output, the sphinx-build must not run in the same process as the server /
sphinx-build must run in a subprocess so that changed source code is
re-imported.

This issue was introduced in commit d57f117 / a commit without a PR & review.

Related: https://github.com/sphinx-doc/sphinx-autobuild/issues/31#issuecomment-111232175
Closes: https://github.com/sphinx-doc/sphinx-autobuild/issues/148

Signed-off-by: Markus Heiser <markus.heiser@darmarit.de>